# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/categories/like:
    put:
      tags:
      - rest-category
      operationId: likeCategory
      parameters:
      - name: categoryName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Category_CategoryDetails'
  /api/categories/edit:
    put:
      tags:
      - rest-category
      summary: Edits an existing category
      operationId: editCategory
      parameters:
      - name: categoryName
        in: query
        required: true
        schema:
          type: string
      - name: newDescription
        in: query
        required: true
        schema:
          type: string
      - name: color
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "404":
          description: Not Found
        "200":
          description: Returns the created category information
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDetails_CategoryDetails'
        "403":
          description: Only access to admin account
        "409":
          description: There is already a category with the name given by parameter
  /api/categories/dislike:
    put:
      tags:
      - rest-category
      operationId: dislikeCategory
      parameters:
      - name: categoryName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Category_CategoryDetails'
...
  /api/categories/new:
    post:
      tags:
      - rest-category
      summary: "Create a category with a specific name, description, color and an\
        \ optional image"
      operationId: createCategory
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: des
        in: query
        required: true
        schema:
          type: string
      - name: color
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "404":
          description: Not Found
        "200":
          description: Returns the created category information
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDetails_CategoryDetails'
        "403":
          description: Only access to admin account
        "409":
          description: There is already a category with the name given by parameter
...
  /api/plans/category:
    get:
      tags:
      - rest-plan
      operationId: getPlansbyCategoryName
      parameters:
      - name: category
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan_PlanDetails'
...
  /api/categories:
    get:
      tags:
      - rest-category
      summary: Get the information a specific category searched by its name
      operationId: categoryInfo
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Doesn't exists a category with that name
          content: {}
        "200":
          description: Returns the category information
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDetails_CategoryDetails'
  /api/categories/:
    get:
      tags:
      - rest-category
      summary: Get the information of all existing categories
      operationId: getCategories
      responses:
        "404":
          description: Not Found
        "200":
          description: Returns all the existing categories
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDetails_CategoryDetails'
...
components:
  schemas:
...
    Category_PlanDetails:
      type: object
      properties:
        name:
          type: string
        des:
          type: string
        color:
          type: string
        likedByUser:
          type: boolean
        plans:
          type: array
          items:
            $ref: '#/components/schemas/Plan_PlanDetails'
        userFavoritesCategory:
          type: array
          items:
            $ref: '#/components/schemas/User_PlanDetails'
        description:
          type: string
    Plan_PlanDetails:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: integer
          format: int32
        likedUser:
          type: boolean
        category:
          $ref: '#/components/schemas/Category_PlanDetails'
    Tree_PlanDetails:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User_PlanDetails'
        category:
          $ref: '#/components/schemas/Category_PlanDetails'
        last_update:
          type: integer
          format: int64
        date:
          type: string
        imagePath:
          type: string
    User_PlanDetails:
      type: object
      properties:
        trees:
          type: array
          items:
            $ref: '#/components/schemas/Tree_PlanDetails'
        encodedPassword:
          type: string
        confirmEncodedPassword:
          type: string
        password:
          type: string
          writeOnly: true
    Category_CategoryDetails:
      type: object
      properties:
        name:
          type: string
        des:
          type: string
        color:
          type: string
        likedByUser:
          type: boolean
        plans:
          type: array
          items:
            $ref: '#/components/schemas/Plan_CategoryDetails'
        userFavoritesCategory:
          type: array
          items:
            $ref: '#/components/schemas/User_CategoryDetails'
        description:
          type: string
    Plan_CategoryDetails:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: integer
          format: int32
        likedUser:
          type: boolean
    TreePK_CategoryDetails:
      type: object
      properties:
        userPK:
          type: string
        categoryPK:
          type: string
    Tree_CategoryDetails:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User_CategoryDetails'
        category:
          $ref: '#/components/schemas/Category_CategoryDetails'
        treePK:
          $ref: '#/components/schemas/TreePK_CategoryDetails'
        height:
          type: integer
          format: int32
        last_update:
          type: integer
          format: int64
        date:
          type: string
        imagePath:
          type: string
    User_CategoryDetails:
      type: object
      properties:
        trees:
          type: array
          items:
            $ref: '#/components/schemas/Tree_CategoryDetails'
        encodedPassword:
          type: string
        confirmEncodedPassword:
          type: string
        password:
          type: string
          writeOnly: true
    CategoryDetails_CategoryDetails:
      type: object
...